// frontend/src/types/api/listSummariesApiSummariesGetSort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListSummariesApiSummariesGetSort = typeof ListSummariesApiSummariesGetSort[keyof typeof ListSummariesApiSummariesGetSort];


export const ListSummariesApiSummariesGetSort = {
  newest: 'newest',
  oldest: 'oldest',
} as const;